% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_model.R
\name{fcn_model}
\alias{fcn_model}
\title{FCN model}
\usage{
fcn_model(input_size, output_size)
}
\arguments{
\item{input_size}{input size (nr of columns with time series drivers)}

\item{output_size}{the output size of the model}
}
\value{
a bare FCN (DNN) model, to be configured and run
}
\description{
FCN model
}
