% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_dataset.R
\name{gpp_dataset}
\alias{gpp_dataset}
\title{Data format routine, queried by data loader}
\usage{
gpp_dataset(x, train_center)
}
\arguments{
\item{x}{input dataset}

\item{train_center}{mean and sd to center the data}
}
\value{
torch tensor for model training
}
\description{
Data format routine, queried by data loader
}
