% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_model.R, R/rnn_model.R
\name{rnn_model}
\alias{rnn_model}
\title{FCN model}
\usage{
rnn_model(input_size, hidden_size, output_size, num_layers = 2, dropout = 0.3)

rnn_model(input_size, hidden_size, output_size, num_layers = 2, dropout = 0.3)
}
\arguments{
\item{input_size}{input size (nr of columns with time series drivers)}

\item{hidden_size}{the size of the LSTM hidden layers in the module}

\item{output_size}{the output size of the model}

\item{num_layers}{number of layers in the LSTM module}

\item{dropout}{dropout rate of the LSTM module}
}
\value{
a bare FCN (DNN) model, to be configured and run

a bare LSTM model, to be configured and run
}
\description{
FCN model

LSTM model
}
